pyproject.toml
[project]
dependencies = ["ruamel.yaml"]
name = "check_gha_pinning"
requires-python = ">= 3.10"
version = "v1.1.0"

[project.optional-dependencies]
test = ["pytest", "pytest-cov"]

[project.scripts]
check-gha-pinning = "check_gha_pinning:main"

[tool.pytest.ini_options]
python_files = "test_*.py"

[tool.ruff]
line-length = 120
